Add files via upload
This Python script serves as an eligibility checker for students who have completed a diploma and are considering pursuing a B.Tech program. It calculates the total credits earned over multiple semesters and determines whether the student meets the eligibility criteria for B.Tech admission. It provides clear feedback based on the total credits, informing the user of their eligibility status. If the total credits meet the required threshold, it congratulates the student, otherwise, it encourages them to try again in the future.
The script is a useful tool for students to quickly assess their eligibility for B.Tech programs and make informed decisions about their academic journey.